scraper/requirements.txt
# Telegram Scraper
telethon>=1.34.0
python-dotenv>=1.0.0
aiohttp>=3.9.0
Pillow>=10.0.0

# OCR
paddlepaddle>=2.6.0
paddleocr>=2.7.0

# Database
asyncpg>=0.29.0
sqlalchemy>=2.0.0

# Utils
python-dateutil>=2.8.0
pytz>=2024.1
